seperated testing functions into two seperate scripts
